requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
firebase-admin==6.3.0
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
numpy==1.24.3
email-validator==2.1.0.post1
gunicorn==21.2.0
werkzeug==3.0.1
